+++ resvg-cairo/examples/minimal.rs
@@ -1,3 +1,5 @@
+use usvg::SystemFontDB;
+
 fn main() {
     let args: Vec<String> = std::env::args().collect();
     if args.len() != 3 {
@@ -7,6 +9,7 @@ fn main() {
 
     let mut opt = usvg::Options::default();
     opt.path = Some(args[1].clone().into());
+    opt.fontdb.load_system_fonts();
 
     let rtree = usvg::Tree::from_file(&args[1], &opt).unwrap();
     let img = resvg_cairo::render_to_image(&rtree, usvg::FitTo::Original, None).unwrap();